# apps/docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenPacks Duel API
  version: 0.2.0-devnet
  summary: Build pack-opening duels on Solana
  description: |
    Preview contract for pack discovery, duel matchmaking, wallet-signed Solana
    transactions, status tracking, and social sharing. The API never accepts
    wallet private keys and does not replace on-chain verification. Every duel
    created by this contract is explicitly Solana devnet-only. Mock provider
    responses never imply funding, settlement, or mainnet readiness.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: OpenPacks Duel
    url: https://github.com/openpacksduel
servers:
  - url: http://localhost:3003/v1
    description: Local development
  - url: https://api.openpacksduel.com/v1
    description: Reserved devnet API endpoint (not live; not mainnet)
...
security: []
paths:
...
  /duels:
    get:
      operationId: listDuels
      tags: [Duels]
      summary: List public duels
      x-openpacksduel-availability: devnet
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DuelStatus'
        - name: wallet
          in: query
          schema:
            $ref: '#/components/schemas/SolanaAddress'
        - name: matchmakingMode
          in: query
          schema:
            $ref: '#/components/schemas/MatchmakingMode'
        - name: packId
          in: query
          schema:
            $ref: '#/components/schemas/PackId'
      responses:
        '200':
          description: Duels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuelList'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      operationId: createDuel
      tags: [Duels]
      summary: Create an off-chain duel intent
      description: The returned intent is not funded until its Solana transaction confirms.
      x-openpacksduel-availability: devnet
      security:
        - walletSession: []
        - integrationKey: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDuelRequest'
      responses:
        '201':
          description: Duel intent created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Duel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /duels/{duelId}:
    get:
      operationId: getDuel
      tags: [Duels]
      summary: Get canonical duel state
      x-openpacksduel-availability: devnet
      parameters:
        - $ref: '#/components/parameters/DuelId'
      responses:
        '200':
          description: Duel state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Duel'
        '404':
          $ref: '#/components/responses/NotFound'
  /duels/{duelId}/receipt:
    get:
      operationId: getPublicDuelReceipt
      tags: [Duels]
      summary: Download the strict public proof for one duel
      description: |
        Returns a no-store, machine-readable snapshot built only from durable
        public duel state. Private transaction metadata, provider errors,
        credentials, and support notes are never included. Missing settlement
        or custody evidence is declared in `availability` instead of inferred.
      x-openpacksduel-availability: devnet
      parameters:
        - $ref: '#/components/parameters/DuelId'
      responses:
        '200':
          description: Downloadable public duel receipt
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="duel_example.receipt.json"
            X-Robots-Tag:
              schema:
                type: string
                const: noindex, nofollow, noarchive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicDuelReceipt'
        '404':
          $ref: '#/components/responses/NotFound'
  /duels/{duelId}/join:
    post:
      operationId: joinDuel
      tags: [Duels]
      summary: Join a direct invitation or open devnet duel
      description: |
        Atomically claims an open duel or accepts a direct invitation. House
        duels are already matched at creation and cannot be joined.
      x-openpacksduel-availability: devnet
      security:
        - walletSession: []
        - integrationKey: []
      parameters:
        - $ref: '#/components/parameters/DuelId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinDuelRequest'
      responses:
        '200':
          description: Duel matched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Duel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /duels/{duelId}/cancel:
    post:
      operationId: cancelDuel
      tags: [Duels]
      summary: Cancel an unfunded devnet duel
      description: |
        A participant can cancel only while the duel is waiting or matched.
        Expired waiting and matched duels deterministically become cancelled
        with the `timeout` reason.
      x-openpacksduel-availability: devnet
      security:
        - walletSession: []
        - integrationKey: []
      parameters:
        - $ref: '#/components/parameters/DuelId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelDuelRequest'
      responses:
        '200':
          description: Duel cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Duel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /duels/{duelId}/events:
    get:
      operationId: listDuelEvents
      tags: [Duels]
      summary: List the authenticated immutable duel state timeline
      description: Internal event data is never exposed by an unauthenticated endpoint.
      x-openpacksduel-availability: devnet
      security:
        - integrationKey: []
      parameters:
        - $ref: '#/components/parameters/DuelId'
      responses:
        '200':
          description: Ordered audit events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DuelEvent'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /duels/{duelId}/transactions:
    get:
      operationId: listDuelTransactions
      tags: [Transactions]
      summary: List authenticated reconciliation-safe Solana transaction records
      description: Raw provider errors and transaction metadata require an integration key.
      x-openpacksduel-availability: devnet
      security:
        - integrationKey: []
      parameters:
        - $ref: '#/components/parameters/DuelId'
      responses:
        '200':
          description: Ordered transaction records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DuelTransactionRecord'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: prepareDuelTransaction
      tags: [Transactions]
      summary: Prepare a wallet-signed Solana transaction
      description: |
        Returns an unsigned devnet transaction that wraps and escrows only the configured
        per-side platform fee. The pack purchase is not included. The API never signs for a wallet.
      x-openpacksduel-availability: devnet
      security:
        - walletSession: []
        - integrationKey: []
      parameters:
        - $ref: '#/components/parameters/DuelId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrepareTransactionRequest'
      responses:
        '201':
          description: Transaction ready for wallet inspection and signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreparedTransaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /duels/{duelId}/social-card:
    get:
      operationId: getDuelSocialCard
      tags: [Social]
      summary: Get canonical share and social-card URLs
      x-openpacksduel-availability: devnet
      parameters:
        - $ref: '#/components/parameters/DuelId'
      responses:
        '200':
          description: Canonical social URLs for the current duel status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SocialCard'
        '404':
          $ref: '#/components/responses/NotFound'
  /duels/{duelId}/open-packs:
    post:
      operationId: openDuelPacks
      tags: [Duels]
      summary: Generate, open, and normalize both funded duel packs
      description: |
        Idempotent server-side orchestration for a funded duel. A wallet
        participant or integration may advance the OpenPacks devnet provider,
        which mints two single-supply SPL demo cards, deposits both in escrow,
        commits the result, and settles the winner on Solana devnet.
      x-openpacksduel-availability: devnet
      security:
        - walletSession: []
        - integrationKey: []
      parameters:
        - $ref: '#/components/parameters/DuelId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Both results are normalized and ready for settlement orchestration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Duel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/Unavailable'
        '503':
          $ref: '#/components/responses/Unavailable'
  /profiles/{wallet}:
    get:
      operationId: getPublicWalletProfile
      tags: [Duels]
      summary: Get a bounded pseudonymous wallet duel summary
      description: |
        Returns recent durable duel activity, record, refunds, and biggest
        recorded win. The wallet remains a public Solana address, while display
        labels are shortened and consumers must not index this response.
      x-openpacksduel-availability: devnet
      parameters:
        - $ref: '#/components/parameters/Wallet'
      responses:
        '200':
          description: Pseudonymous wallet profile
          headers:
            X-Robots-Tag:
              schema:
                type: string
                const: noindex, nofollow, noarchive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicWalletProfile'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cronSecret:
      type: http
      scheme: bearer
      bearerFormat: random server secret
      description: Vercel Cron bearer secret; accepted only by internal worker routes.
    integrationKey:
      type: http
      scheme: bearer
      bearerFormat: opd_live_...
      description: Server-side integration key. Never expose it in browser code.
    walletSession:
      type: http
      scheme: bearer
      bearerFormat: opd_devnet_session_...
      description: Opaque 15-minute session issued after a Solana devnet Ed25519 signature.
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 256
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    DuelId:
      name: duelId
      in: path
      required: true
      schema:
        type: string
        pattern: '^duel_[A-Za-z0-9]{12,64}$'
    Wallet:
      name: wallet
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/SolanaAddress'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing, invalid, or expired bearer credential
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Request conflicts with current duel state
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unavailable:
      description: Required server configuration or dependency is unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    SolanaAddress:
      type: string
      pattern: '^[1-9A-HJ-NP-Za-km-z]{32,44}$'
      examples: [9xQeWvG816bUx9EPfEZvD6nGQ3xM4wzHY6zvQ3z9gJ1]
    PackId:
      type: string
      pattern: '^[a-z0-9][a-z0-9_-]{2,63}$'
    Money:
      type: object
      additionalProperties: false
      required: [amount, currency, decimals]
      properties:
        amount:
          type: string
          pattern: '^[0-9]+$'
          description: Integer minor units encoded as a string.
        currency:
          type: string
          enum: [USDC]
        decimals:
          type: integer
          const: 6
...
    MatchmakingMode:
      type: string
      enum: [direct, house, open]
...
    DuelStatus:
      type: string
      enum:
        - waiting
        - matched
        - committing
        - funded
        - opening
        - awaiting_assets
        - settling
        - settled
        - cancelling
        - cancelled
        - refunding
        - refunded
        - failed
    CreateDuelRequest:
      type: object
      additionalProperties: false
      required: [creatorWallet, packId, matchmakingMode, expiresAt]
      properties:
        analyticsSessionId:
          type: string
          pattern: '^anon_[a-f0-9]{32}$'
        creatorWallet:
          $ref: '#/components/schemas/SolanaAddress'
        opponentWallet:
          oneOf:
            - $ref: '#/components/schemas/SolanaAddress'
            - type: 'null'
        packId:
          $ref: '#/components/schemas/PackId'
        matchmakingMode:
          $ref: '#/components/schemas/MatchmakingMode'
        expiresAt:
          type: string
          format: date-time
          description: Must be in the future and no more than 24 hours from creation.
    JoinDuelRequest:
      type: object
      additionalProperties: false
      required: [wallet]
      properties:
        analyticsSessionId:
          type: string
          pattern: '^anon_[a-f0-9]{32}$'
        wallet:
          $ref: '#/components/schemas/SolanaAddress'
    CancelDuelRequest:
      type: object
      additionalProperties: false
      required: [wallet]
      properties:
        analyticsSessionId:
          type: string
          pattern: '^anon_[a-f0-9]{32}$'
        wallet:
          $ref: '#/components/schemas/SolanaAddress'
        reason:
          type: string
          maxLength: 160
    Duel:
      type: object
      additionalProperties: false
      required:
        - id
        - status
        - matchmakingMode
        - creatorWallet
        - environment
        - providerMode
        - houseOpponent
        - pack
        - stake
        - version
        - createdAt
        - expiresAt
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/DuelStatus'
        matchmakingMode:
          $ref: '#/components/schemas/MatchmakingMode'
        creatorWallet:
          $ref: '#/components/schemas/SolanaAddress'
        environment:
          type: string
          const: solana-devnet
        providerMode:
          type: string
          enum: [mock, openpacksduel-devnet, collector-crypt-sandbox]
          description: Mock outcomes are valueless and never represent transferred assets.
        houseOpponent:
          type: boolean
        opponentWallet:
          oneOf:
            - $ref: '#/components/schemas/SolanaAddress'
            - type: 'null'
        pack:
          $ref: '#/components/schemas/Pack'
        stake:
          $ref: '#/components/schemas/Money'
        escrowAddress:
          oneOf:
            - $ref: '#/components/schemas/SolanaAddress'
            - type: 'null'
        transactionSignature:
          type: [string, 'null']
        winnerWallet:
          oneOf:
            - $ref: '#/components/schemas/SolanaAddress'
            - type: 'null'
        result:
          oneOf:
            - $ref: '#/components/schemas/DuelResult'
            - type: 'null'
        opponentJoinedAt:
          type: [string, 'null']
          format: date-time
        cancellationReason:
          type: [string, 'null']
        version:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PublicDuelStatus:
      oneOf:
        - $ref: '#/components/schemas/DuelStatus'
        - type: string
          const: expired
...
    PublicDuelReceipt:
      type: object
      additionalProperties: false
      required: [schemaVersion, duel, pack, participants, result, fees, custody, references, recovery, availability, cardActions, actions, privacy]
      properties:
        schemaVersion:
          type: string
          const: openpacksduel.receipt.v1
        duel:
          type: object
          additionalProperties: false
          required: [id, status, mode, network, createdAt, expiresAt, observedAt]
          properties:
            id: { type: string }
            status: { $ref: '#/components/schemas/PublicDuelStatus' }
            mode: { $ref: '#/components/schemas/MatchmakingMode' }
            network: { type: string, const: solana-devnet }
            createdAt: { type: string, format: date-time }
            expiresAt: { type: string, format: date-time }
            observedAt: { type: string, format: date-time }
        pack:
          type: object
          additionalProperties: false
          required: [id, name, provider, providerMode, providerPackId, tier]
          properties:
            id: { type: string }
            name: { type: string }
            provider: { type: string }
            providerMode:
              type: string
              enum: [mock, openpacksduel-devnet, collector-crypt-sandbox]
            providerPackId: { type: [string, 'null'] }
            tier: { $ref: '#/components/schemas/Money' }
        participants:
          type: object
          additionalProperties: false
          required: [creator, opponent]
          properties:
            creator: { $ref: '#/components/schemas/PublicParticipant' }
            opponent:
              oneOf:
                - $ref: '#/components/schemas/PublicParticipant'
                - type: 'null'
        result:
          oneOf:
            - $ref: '#/components/schemas/PublicDuelResult'
            - type: 'null'
        fees:
          type: object
          additionalProperties: false
          required: [asset, perSideAmountLamports, finalizedSides, requiredSides, totalFinalizedAmountLamports]
          properties:
            asset: { type: string, const: WSOL }
            perSideAmountLamports: { type: [string, 'null'], pattern: '^\d+$' }
            finalizedSides: { type: integer, minimum: 0, maximum: 2 }
            requiredSides: { type: integer, const: 2 }
            totalFinalizedAmountLamports: { type: [string, 'null'], pattern: '^\d+$' }
        custody:
          type: object
          additionalProperties: false
          required: [platformFee, cardAssets]
          properties:
            platformFee:
              type: object
              additionalProperties: false
              required: [asset, escrowAddress, status]
              properties:
                asset: { type: string, const: WSOL }
                escrowAddress:
                  oneOf:
                    - $ref: '#/components/schemas/SolanaAddress'
                    - type: 'null'
                status: { type: string, enum: [not-started, partially-funded, funded] }
            cardAssets:
              type: object
              additionalProperties: false
              required: [status, detail]
              properties:
                status: { type: string, enum: [not-recorded, provider-results-recorded, settlement-reference-recorded] }
                detail: { type: string }
        references:
          type: object
          additionalProperties: false
          required: [provider, solana]
          properties:
            provider:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [side, provider, providerReference, assetReference]
                properties:
                  side: { type: string, enum: [creator, opponent] }
                  provider: { type: string }
                  providerReference: { type: string }
                  assetReference: { type: string }
            solana:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [action, bindingSource, signature, status, finalizedAt, recoveredAt, explorerUrl]
                properties:
                  action:
                    type: string
                    enum: [fund, cancel, refund, open_pack, commit_result, settle]
                  bindingSource: { type: string, enum: [api-submission, rpc-recovery] }
                  signature: { type: string }
                  status: { type: string, enum: [prepared, submitted, confirmed, finalized, failed, expired] }
                  finalizedAt: { type: [string, 'null'], format: date-time }
                  recoveredAt: { type: [string, 'null'], format: date-time }
                  explorerUrl: { type: string, format: uri }
        recovery:
          type: object
          additionalProperties: false
          required: [status, alerts]
          properties:
            status: { type: string, enum: [none, recovered, attention-required] }
            alerts:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [action, code, detectedAt, signature, explorerUrl]
                properties:
                  action: { type: string, const: fund }
                  code: { type: string, const: UNBOUND_FINALIZED_ESCROW_STATE_MISMATCH }
                  detectedAt: { type: string, format: date-time }
                  signature: { type: string }
                  explorerUrl: { type: string, format: uri }
        availability:
          type: object
          additionalProperties: false
          required: [complete, missing]
          properties:
            complete: { type: boolean }
            missing: { type: array, items: { type: string } }
        cardActions:
          $ref: '#/components/schemas/PublicPostDuelCardActions'
        actions:
          $ref: '#/components/schemas/PublicActions'
        privacy:
          $ref: '#/components/schemas/PublicPrivacy'
...
    DuelList:
      type: object
      additionalProperties: false
      required: [data, hasMore]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Duel'
        nextCursor:
          type: [string, 'null']
        hasMore:
          type: boolean
    DuelEvent:
      type: object
      additionalProperties: false
      required: [id, duelId, sequence, type, createdAt]
      properties:
        id:
          type: string
        duelId:
          type: string
        sequence:
          type: integer
          minimum: 1
        type:
          type: string
        actorWallet:
          oneOf:
            - $ref: '#/components/schemas/SolanaAddress'
            - type: 'null'
        fromStatus:
          oneOf:
            - $ref: '#/components/schemas/DuelStatus'
            - type: 'null'
        toStatus:
          oneOf:
            - $ref: '#/components/schemas/DuelStatus'
            - type: 'null'
        data:
          type: [object, 'null']
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    PrepareTransactionRequest:
      type: object
      additionalProperties: false
      required: [action, wallet]
      properties:
        action:
          type: string
          const: fund
        wallet:
          $ref: '#/components/schemas/SolanaAddress'
...
    DuelTransactionRecord:
      type: object
      additionalProperties: false
      required: [id, duelId, action, status, network, wallet, createdAt, updatedAt]
      properties:
        id:
          type: string
        duelId:
          type: string
        action:
          type: string
          enum: [fund, cancel, refund, open_pack, commit_result, settle]
        status:
          type: string
          enum: [prepared, submitted, confirmed, finalized, failed, expired]
        network:
          type: string
          const: solana-devnet
        wallet:
          $ref: '#/components/schemas/SolanaAddress'
        signature:
          type: [string, 'null']
        recentBlockhash:
          type: [string, 'null']
        recoveredAt:
          type: [string, 'null']
          format: date-time
          description: Set only after bounded escrow-PDA recovery verifies a persisted prepared intent.
        recoveryAlertCode:
          type: [string, 'null']
          enum: [UNBOUND_FINALIZED_ESCROW_STATE_MISMATCH, null]
        recoveryCandidateAt:
          type: [string, 'null']
          format: date-time
        recoveryCandidateSignature:
          type: [string, 'null']
        lastValidBlockHeight:
          type: [string, 'null']
        providerReference:
          type: [string, 'null']
        errorCode:
          type: [string, 'null']
        errorMessage:
          type: [string, 'null']
        submittedAt:
          type: [string, 'null']
          format: date-time
        confirmedAt:
          type: [string, 'null']
          format: date-time
        finalizedAt:
          type: [string, 'null']
          format: date-time
        confirmationStatus:
          type: [string, 'null']
          enum: [confirmed, finalized, null]
        checkAttempts:
          type: integer
          minimum: 0
        lastCheckedAt:
          type: [string, 'null']
          format: date-time
        stuckAt:
          type: [string, 'null']
          format: date-time
        expiresAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SocialCard:
      type: object
      additionalProperties: false
      required: [duelId, status, pageUrl, imageUrl]
      properties:
        duelId:
          type: string
        status:
          $ref: '#/components/schemas/DuelStatus'
        pageUrl:
          type: string
          format: uri
        imageUrl:
          type: string
          format: uri
        shareText:
          type: string
          maxLength: 280
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, requestId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        requestId:
          type: string